requirements.txt
aiofiles==0.7.0
aiohttp==3.7.4.post0
alexapy==1.25.1
appdirs==1.4.4
async-timeout==3.0.1
attrs==21.2.0
authcaptureproxy==1.0.1
backoff==1.10.0
backtrader==1.9.76.123
beautifulsoup4==4.9.3
bs4==0.0.1
CacheControl==0.12.6
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
dateparser==1.0.0
distlib==0.3.2
filelock==3.0.12
firebase-admin==5.0.0
gevent==21.1.2
google-api-core==1.30.0
google-api-python-client==2.8.0
google-auth==1.31.0
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.1.1
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
greenlet==1.1.0
grpcio==1.38.0
h11==0.12.0
httpcore==0.13.3
httplib2==0.19.1
httpx==0.18.1
idna==2.10
importlib-metadata==3.10.1
msgpack==1.0.2
multidict==5.1.0
numpy==1.20.3
packaging==20.9
proto-plus==1.18.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pync==2.0.3
pyotp==2.6.0
pyparsing==2.4.7
python-binance==1.0.12
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
rfc3986==1.5.0
rsa==4.7.2
simplejson==3.17.2
six==1.16.0
sniffio==1.2.0
soupsieve==2.2.1
TA-Lib==0.4.20
typer==0.3.2
typing-extensions==3.10.0.0
tzlocal==2.1
ujson==4.0.2
uritemplate==3.0.1
urllib3==1.26.5
virtualenv==20.4.7
websocket-client==1.0.1
websockets==9.1
yarl==1.6.3
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.4.0
